fix: add matchAndTapImage to DeviceWrapper
