+++ package.json
@@ -62,6 +62,7 @@
     "@appium/support@npm:^5.1.6": "patch:@appium/support@npm%3A5.1.2#~/patches/@appium-support-npm-5.1.2-0c5ea57d71.patch",
     "@appium/support@npm:^5.1.2": "patch:@appium/support@npm%3A5.1.2#~/patches/@appium-support-npm-5.1.2-0c5ea57d71.patch",
     "@appium/support@npm:^5.0.3": "patch:@appium/support@npm%3A5.1.2#~/patches/@appium-support-npm-5.1.2-0c5ea57d71.patch",
-    "@appium/support@npm:^5.1.1": "patch:@appium/support@npm%3A5.1.2#~/patches/@appium-support-npm-5.1.2-0c5ea57d71.patch"
+    "@appium/support@npm:^5.1.1": "patch:@appium/support@npm%3A5.1.2#~/patches/@appium-support-npm-5.1.2-0c5ea57d71.patch",
+    "sharp": "^0.34.1"
   }
 }
